% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer_style.R
\name{new_beamer_style}
\alias{new_beamer_style}
\title{Create beamer style-sheet
This function creates an 'assets' folder at the working directory.
The folder contains a stylesheet and other auxiliary files to be applied
to an ioslide-document.}
\usage{
new_beamer_style()
}
\description{
Create beamer style-sheet
This function creates an 'assets' folder at the working directory.
The folder contains a stylesheet and other auxiliary files to be applied
to an ioslide-document.
}
