% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer_style.R
\name{torino}
\alias{torino}
\title{torino-style ioslides}
\usage{
torino(...)
}
\arguments{
\item{...}{Extra parameters to pass to `rmarkdown::ioslides_presentation`.}
}
\description{
torino-style ioslides
}
