% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer_style.R
\name{TUDelft}
\alias{TUDelft}
\title{TUDelft-style ioslides}
\usage{
TUDelft(...)
}
\arguments{
\item{...}{Extra parameters to pass to `rmarkdown::ioslides_presentation`.}
}
\description{
TUDelft-style ioslides
}
\references{
https://www.tudelft.nl/huisstijl/kleuren/
}
