% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer_style.R
\name{torino}
\alias{torino}
\title{torino-style ioslides}
\usage{
torino(...)
}
\arguments{
\item{...}{Extra parameters to pass to `rmarkdown::ioslides_presentation`.}
}
\description{
torino-style ioslides
}
\note{
The original torino style is designed by Marco Barisione, see references for more details.
}
\references{
https://github.com/bbatsov/beamer-torino-theme

https://blog.barisione.org/2007/09/18/torino-a-pretty-theme-for-latex-beamer/
}
